RHINENG-23032: Split app.models.schemas into package modules
Replace monolithic schemas.py with app/models/schemas/: common base,
system_profile, host, mq, staleness, outbox, and consumers/* for host app
data schemas. Introduce BaseSchemaWithExclude for repeated Meta.unknown=EXCLUDE.

Point system_profile_normalizer at app.models.schemas.system_profile to avoid
circular imports. build_model uses app.models.schemas Host/LimitedHost so
unit test patches keep working.

Made-with: Cursor
