+++ app/models/schemas/_log.py
@@ -1,0 +1,3 @@
+from app.logging import get_logger
+
+logger = get_logger("app.models.schemas")
