+++ app/models/schemas/consumers/malware.py
@@ -1,0 +1,12 @@
+from marshmallow import Schema as MarshmallowSchema
+from marshmallow import fields
+from marshmallow import validate as marshmallow_validate
+
+
+class MalwareDataSchema(MarshmallowSchema):
+    """Schema for Malware application data."""
+
+    last_status = fields.Str(allow_none=True, validate=marshmallow_validate.Length(max=50))
+    last_matches = fields.Int(allow_none=True)
+    total_matches = fields.Int(allow_none=True)
+    last_scan = fields.DateTime(allow_none=True)
