+++ app/models/schemas/consumers/vulnerability.py
@@ -1,0 +1,12 @@
+from marshmallow import Schema as MarshmallowSchema
+from marshmallow import fields
+
+
+class VulnerabilityDataSchema(MarshmallowSchema):
+    """Schema for Vulnerability application data."""
+
+    total_cves = fields.Int(allow_none=True)
+    critical_cves = fields.Int(allow_none=True)
+    high_severity_cves = fields.Int(allow_none=True)
+    cves_with_security_rules = fields.Int(allow_none=True)
+    cves_with_known_exploits = fields.Int(allow_none=True)
