+++ app/models/schemas/system_profile.py
@@ -1,0 +1,63 @@
+from marshmallow import Schema as MarshmallowSchema
+from marshmallow import fields
+from marshmallow import validate as marshmallow_validate
+
+from app.models.system_profile_normalizer import SystemProfileNormalizer
+from app.validators import check_empty_keys
+
+
+# Nested field schemas used in system profile JSON; plain Marshmallow, no EXCLUDE.
+class DiskDeviceSchema(MarshmallowSchema):
+    device = fields.Str(validate=marshmallow_validate.Length(max=2048))
+    label = fields.Str(validate=marshmallow_validate.Length(max=1024))
+    options = fields.Dict(validate=check_empty_keys)
+    mount_point = fields.Str(validate=marshmallow_validate.Length(max=2048))
+    type = fields.Str(validate=marshmallow_validate.Length(max=256))
+
+
+class RhsmSchema(MarshmallowSchema):
+    version = fields.Str(validate=marshmallow_validate.Length(max=256))
+    environment_ids = fields.List(fields.Str(validate=marshmallow_validate.Length(max=256)))
+
+
+class OperatingSystemSchema(MarshmallowSchema):
+    major = fields.Int()
+    minor = fields.Int()
+    name = fields.Str(validate=marshmallow_validate.Length(max=256))
+
+
+class YumRepoSchema(MarshmallowSchema):
+    id = fields.Str(validate=marshmallow_validate.Length(max=256))
+    name = fields.Str(validate=marshmallow_validate.Length(max=1024))
+    gpgcheck = fields.Bool()
+    enabled = fields.Bool()
+    base_url = fields.Str(validate=marshmallow_validate.Length(max=2048))
+    mirrorlist = fields.Str(validate=marshmallow_validate.Length(max=2048))
+
+
+class DnfModuleSchema(MarshmallowSchema):
+    name = fields.Str(validate=marshmallow_validate.Length(max=128))
+    stream = fields.Str(validate=marshmallow_validate.Length(max=2048))
+    status = fields.List(fields.Str(validate=marshmallow_validate.Length(max=64)))
+
+
+class InstalledProductSchema(MarshmallowSchema):
+    name = fields.Str(validate=marshmallow_validate.Length(max=512))
+    id = fields.Str(validate=marshmallow_validate.Length(max=64))
+    status = fields.Str(validate=marshmallow_validate.Length(max=256))
+
+
+class NetworkInterfaceSchema(MarshmallowSchema):
+    ipv4_addresses = fields.List(fields.Str())
+    ipv6_addresses = fields.List(fields.Str())
+    state = fields.Str(validate=marshmallow_validate.Length(max=25))
+    mtu = fields.Int()
+    mac_address = fields.Str(validate=marshmallow_validate.Length(max=59))
+    name = fields.Str(validate=marshmallow_validate.Length(min=1, max=50))
+    state = fields.Str(validate=marshmallow_validate.Length(max=25))
+    type = fields.Str(validate=marshmallow_validate.Length(max=18))
+
+
+_normalizer = SystemProfileNormalizer()
+HostStaticSystemProfileSchema = _normalizer.create_static_schema()
+HostDynamicSystemProfileSchema = _normalizer.create_dynamic_schema()
